+++ ydb/apps/ydb/CHANGELOG.md
@@ -1,3 +1,4 @@
+* Added support of check canonical results in `ydb workload query run` command.
 * Added support of parquet format to `ydb workload query import` and `ydb workload clickbench import files` commands.
 * `ydb workload query` deep reworked. Now it can be used for user defined full cycle load testing.
 * Fixed an issue where the `ydb workload * clean` commands were deleting all contents from the target directory, instead of just the tables created by the init command.